Refactor to split obstree.
