+++ gsadjust/models/__init__.py
@@ -3,13 +3,6 @@ from .calibration import MeterCalibrationModel
 from .datum import DatumTableModel
 from .delta import DeltaTableModel, NoCheckDeltaTableModel
 from .gravity import GravityChangeModel
-from .obstree import (
-    ObsTreeItem,
-    ObsTreeLoop,
-    ObsTreeModel,
-    ObsTreeStation,
-    ObsTreeSurvey,
-)
 from .result import ResultsTableModel
 from .roman import RomanTableModel
 from .scintrex import ScintrexTableModel
