+++ gsadjust/obstree/survey.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-#  -*- coding: utf-8 -*-
 """
 pyqt_modules.py
 ===============
@@ -44,11 +42,9 @@ from ..data import (
 )
 from ..data.analysis import InversionError, numpy_inversion
 from ..gui.messages import show_message
-from .datum import DatumTableModel
-from .delta import DeltaTableModel
-from .result import ResultsTableModel
-from .tare import TareTableModel
-from .utils import format_numeric_column
+from ..models import DatumTableModel, DeltaTableModel, ResultsTableModel
+from .base import ObsTreeItemBase
+from .loop import ObsTreeLoop
 
 # Constants for column headers
 STATION_NAME, STATION_DATETIME, STATION_MEAN = range(3)
@@ -56,609 +52,7 @@ LOOP_NAME = 0
 SURVEY_NAME = 0
 
 
-class tempStation:
-    def __init__(self, station):
-        self.__dict__ = station
-
-
-class ObsTreeItem(QtGui.QStandardItem):
-    """
-    Basic tree-view item used to populate data tree, used for Surveys, Loops, and Stations.
-    Not used directly but inherited by ObsTreeStation, ...Loop, and ...Survey
-    """
-
-    def __init__(self):
-        super(ObsTreeItem, self).__init__()
-        self.setFlags(
-            self.flags() | Qt.ItemIsEnabled | Qt.ItemIsEditable | Qt.ItemIsUserCheckable
-        )
-
-        self.setCheckState(Qt.Checked)
-        self.fontweight = QtGui.QFont.Normal
-        self.cellcolor = Qt.white
-
-
-class ObsTreeStation(ObsTreeItem):
-    """
-    PyQt model for stations. Station data is stored as a Station object in .station
-    """
-
-    def __init__(self, k, station_name, station_count):
-        """
-        Create a new station from ChannelList object. The fields (lists) of the ChannelList are copies to the
-        ObsTreeStation.
-        :param k: ChannelList object.
-        :param station_name: String
-        :param station_count: Int
-        """
-        super(ObsTreeStation, self).__init__()  # call properties from the baseclass
-        self.__dict__ = copy.deepcopy(k.__dict__)
-        self.station_name = station_name
-        self.station_count = station_count
-        if not hasattr(k, 'asd'):
-            asd = None  # Records the number of times the station is occupied in a loop
-        if hasattr(k, 'checked'):
-            self.setCheckState(k.checked)
-        # For legacy .p files
-        # if len(self.corr_g) == 0:
-        #     self.corr_g = self.raw_grav
-
-    def __str__(self):
-        return self.station_name
-
-    def _weights_(self):
-        """
-        weights for Scintrex data are calculated based on the sample standard
-        deviation divided by the duration. Neither of these are reported by the
-        Burris meter.  Instead, duration is set to a constant (5) when the file
-        is read. Parameter sd is calculated from the multiple samples comprising
-        a station. Therefore w is a constant array if Burris data, variable
-        if Scintrex.
-
-        TODO: test effect of weighting if Burris and Scintrex data are combined
-        in a survey
-        :return:
-        """
-        if self.meter_type == 'Burris' or self.meter_type == 'CG6Tsoft':
-            return [1 for i in self.keepdata if i == 1]
-        else:
-            # sdtmp = [self.sd[i] / np.sqrt(self.dur[i]) for i in range(len(self.t))]
-            w = [
-                1.0 / (self.sd[i] * self.sd[i])
-                for i in range(len(self.t))
-                if self.keepdata[i] == 1
-            ]
-            return w
-
-    @classmethod
-    def from_station(cls, station):
-        """
-        Create a station from an existing station. Used to clone stations.
-        :param station: ObsTreeStation object to be copied
-        :return: ObsTreeStation
-        """
-        temp = cls(station, station.station_name, station.station_count)
-        temp.__dict__ = station.__dict__
-        return temp
-
-    @property
-    def key(self):
-        return (self.station_name, self.tmean())
-
-    # @property
-    def grav(self):
-        """
-        Applies tares and earth tide correction to raw_grav
-        :return: List
-        """
-        # if len(self.corr_g) == 0:
-        #     self.corr_g = self.raw_grav
-        # data = np.array(self.raw_grav) - np.array(self.tare) + np.array(self.etc)
-        data = [a - b + c for (a, b, c) in zip(self.raw_grav, self.tare, self.etc)]
-        return data
-
-    def display_name(self):
-        return self.station_name + ' (' + self.station_count + ')'
-
-    def display_datetime_or_tmean(self):
-        return (
-            num2date(self.tmean()).strftime('%Y-%m-%d %H:%M:%S')
-            if self.tmean() != -999
-            else self.tmean()
-        )
-
-    def display_mean_stddev(self):
-        return '{:.1f} ± {:.1f}'.format(self.gmean(), self.stdev())
-
-    @property
-    def display_column_map(self):
-        return {
-            STATION_NAME: (self.display_name,),
-            STATION_DATETIME: (self.display_datetime_or_tmean,),
-            STATION_MEAN: (self.display_mean_stddev,),
-        }
-
-    @property
-    def tooltip(self):
-        return None
-
-    # @property
-    def gmean(self):
-        """
-        The try-except block handles errors when all keepdata == 0.
-        """
-        g = self.grav()
-        try:
-            gtmp = [g[i] for i in range(len(self.t)) if self.keepdata[i] == 1]
-            d = sum(self._weights_())
-            w = self._weights_()
-            wg = [g * w for (g, w) in zip(gtmp, w)]
-            gmean = sum(wg) / d
-            return gmean
-        except:
-            return -999
-
-    # @property
-    def tmean(self):
-        """
-        The try-except block handles errors when all keepdata == 0.
-        """
-        try:
-            ttmp = [self.t[i] for i in range(len(self.t)) if self.keepdata[i] == 1]
-            d = sum(self._weights_())
-            w = self._weights_()
-            wt = [g * w for (g, w) in zip(ttmp, w)]
-            tmean = sum(wt) / d
-            return tmean
-        except:
-            return -999
-
-    @property
-    def t_stdev(self):
-        try:
-            if self.sd[0] == -999:
-                return 1
-            else:
-                ttmp = np.array(
-                    [self.t[i] for i in range(len(self.t)) if self.keepdata[i] == 1]
-                )
-                num = np.zeros(len(ttmp))
-                for i in range(len(ttmp)):
-                    num[i] = self._weights_()[i] * ((ttmp[i] - np.mean(ttmp)) * 24) ** 2
-                sd = np.sqrt(
-                    np.sum(num)
-                    / ((len(self._weights_()) - 1) * np.sum(self._weights_()))
-                )  # np.sqrt(1. / sum(
-                # self._weights_()))
-                return sd
-        except:
-            return -999
-
-    @property
-    def original_sd(self):
-        g = self.grav()
-        sdtmp = np.array(
-            [self.sd[i] for i in range(len(self.t)) if self.keepdata[i] == 1]
-        )
-        gtmp = np.array([g[i] for i in range(len(self.t)) if self.keepdata[i] == 1])
-        num = np.zeros(len(sdtmp))
-        for i in range(len(sdtmp)):
-            num[i] = self._weights_()[i] * (gtmp[i] - np.mean(gtmp)) ** 2
-        sd = np.sqrt(
-            np.sum(num) / ((len(self._weights_()) - 1) * np.sum(self._weights_()))
-        )
-        return sd
-
-    # @property
-    def stdev(self):
-        """
-        The try-except block handles errors when all keepdata == 0.
-
-        The Scintrex meter reports an S.D. for each sample; the Burris meter
-        does not report S.D. (with the Palm PDA, it shows S.D. on the display
-        but does not record it).
-        """
-        try:
-            g = self.grav()
-            if hasattr(self, 'assigned_sd'):
-                if self.assigned_sd:
-                    return self.assigned_sd
-            if (
-                self.sd[0] == -999
-            ):  # Burris meter: return the S.D. calculated from all the samples at a station
-                gtmp = np.array(
-                    [g[i] for i in range(len(self.t)) if self.keepdata[i] == 1]
-                )
-                if len(gtmp) == 1:  # Can't take S.D. if there's only one sample
-                    return 3.0
-                else:
-                    return float(np.std(gtmp))
-            else:  # Scintrex: return the average SD of the samples
-                sdtmp = np.array(
-                    [self.sd[i] for i in range(len(self.t)) if self.keepdata[i] == 1]
-                )
-                # sd = np.mean(sdtmp)  #
-                gtmp = np.array(
-                    [g[i] for i in range(len(self.t)) if self.keepdata[i] == 1]
-                )
-                num = np.zeros(len(sdtmp))
-
-                for i in range(len(sdtmp)):
-                    num[i] = self._weights_()[i] * (gtmp[i] - np.mean(gtmp)) ** 2
-                sd = np.sqrt(
-                    np.sum(num)
-                    / ((len(self._weights_()) - 1) * np.sum(self._weights_()))
-                )  # np.sqrt(1. / sum(
-                return sd
-        except:
-            return -999
-
-    @property
-    def summary_str(self):
-        if self.tmean() == -999:
-            tm = self.tmean()
-        else:
-            tm = num2date(self.tmean()).strftime('%Y-%m-%d %H:%M:%S')
-        if len(self.lat) == 0:
-            self.lat, self.long, self.elev = [0], [0], [0]
-        summary_str = '{} {} {} {} {} {} {} {:0.3f} {:0.3f}\n'.format(
-            self.display_name(),
-            tm,
-            self.oper[0],
-            self.meter_type,
-            self.lat[0],
-            self.long[0],
-            self.elev[0],
-            self.gmean(),
-            self.stdev(),
-        )
-        return summary_str
-
-    def iter_samples(self):
-        """
-        Iterator that returns print statements for each sample, used when
-        writing adjustment summary
-        :return: All of the stations in a campaign
-        """
-        g = self.grav()
-        for i in range(len(self.raw_grav)):
-            yield '{} {} {:0.2f} {:0.2f} {:0.2f} {:0.2f}\n'.format(
-                self.keepdata[i],
-                self.station[i],
-                self.raw_grav[i],
-                self.etc[i],
-                g[i],
-                self.sd[i],
-            )
-
-    def to_json(self):
-        jsonalizable_dict = self.__dict__
-        jsonalizable_dict['checked'] = self.checkState()
-        return self.__dict__
-
-
-class ObsTreeLoop(ObsTreeItem):
-    """
-    PyQt model for Loops, parent item for stations.
-    """
-
-    def __init__(self, name):
-        super(ObsTreeLoop, self).__init__()
-        self.delta_model = DeltaTableModel()
-        self.tare_model = TareTableModel()
-        self.name = name
-        self.drift_method = 'none'  # 'none', 'netadj', 'roman', or 'continuous'
-        self.drift_cont_method = (
-            0  # If continuous model, also need to keep track of which type of model
-        )
-        self.drift_cont_startend = (
-            0  # behavior at start/end. 0: extrapolate, 1: constant
-        )
-        self.drift_cont_weighting = 0  # check box check state
-        self.drift_netadj_method = (
-            2  # If netadj method, keep track of polynomial degree
-        )
-        self.comment = ''  # String that can be specified from GUI
-        # TODO: Import comment from Burris file?
-        self.source = ''  # Filename of raw input file
-
-    def __str__(self):
-        if self.drift_method == 'roman' or self.drift_method == 'none':
-            return (
-                'Loop: {}, '
-                'Drift method: {}, '
-                'Meter type: {}\n'.format(self.name, self.drift_method, self.meter_type)
-            )
-
-        elif self.drift_method == 'continuous':
-            return (
-                'Loop: {}, '
-                'Drift method: {}, '
-                'Meter type: {}, '
-                'Continuous drift method: {}, '
-                'Continuous drift start/end method: {}, '
-                'Weighting: {}\n'.format(
-                    self.name,
-                    self.drift_method,
-                    self.meter_type,
-                    self.drift_cont_method,
-                    self.drift_cont_startend,
-                    self.drift_cont_weighting,
-                )
-            )
-        elif self.drift_method == 'netadj':
-            return (
-                'Loop: {}, '
-                'Drift method: {}, '
-                'Meter type: {}, '
-                'Netadj method: {}\n'.format(
-                    self.name,
-                    self.drift_method,
-                    self.meter_type,
-                    self.drift_netadj_method,
-                )
-            )
-
-    @property
-    def oper(self):
-        try:
-            return self.child(0).oper[0]
-        except AttributeError:
-            return ''
-
-    @property
-    def meter(self):
-        try:
-            return self.child(0).meter[0]
-        except AttributeError:
-            return ''
-
-    @property
-    def tooltip(self):
-        return (
-            'Loop: {}\n'
-            'Drift method: {}\n'
-            'Meter: {}\n'
-            'Operator: {}\n'
-            'Comment: {}\n'
-            'Source: {}'.format(
-                self.name,
-                self.__dict__.get('drift_method', ''),
-                self.__dict__.get('meter', ''),
-                self.__dict__.get('oper', ''),
-                self.__dict__.get('comment', ''),
-                self.__dict__.get('source', ''),
-            )
-        )
-
-    @property
-    def display_column_map(self):
-        return {
-            # Column name map
-            # index: name
-            LOOP_NAME: (str, self.name),
-            1: (str, ''),
-            2: (str, ''),
-        }
-
-    @property
-    def meter_type(self):
-        station = self.child(0)
-        # To deal with old-format files (2020-06). Can probably be deleted someday.
-        if station.meter_type == 'Scintrex':
-            return 'CG5'
-        else:
-            return station.meter_type
-
-    @classmethod
-    def from_json(cls, data):
-        temp = cls(data['name'])
-        temp.__dict__ = data
-        potential_missing_fields = {
-            'comment': '',
-            'oper': '',
-            'source': '',
-            'drift_cont_weighting': 0,
-        }
-        temp.delta_model = DeltaTableModel()
-        temp.tare_model = TareTableModel()
-        for tare_dict in data['tares']:
-            tare_object = Tare(
-                tare_dict['datetime'].date(),
-                tare_dict['datetime'].time(),
-                tare_dict['tare'],
-            )
-            temp.tare_model.insertRows(tare_object, 0)
-        if 'checked' in data:
-            temp.setCheckState(data['checked'])
-        for k, v in potential_missing_fields.items():
-            if not hasattr(temp, k):
-                setattr(temp, k, v)
-        return temp
-
-    def rename(self, from_name, to_name):
-        for i in range(self.rowCount()):
-            item = self.child(i, 0)
-            if item.station_name == from_name:
-                item.station_name = to_name
-                for iiii in range(len(item.station)):
-                    item.station[iiii] = to_name
-
-    def populate(self, data):
-        """
-        Populate loop dictionary using data passed as an option. For now, only
-        the 'single' option works.
-
-        :param data: Passed to Loop.populate_station_dic()
-        """
-        prev_sta = data.station[0]
-        ind_start = 0
-        station_count_dic = dict()
-
-        # loop over samples:
-        for i in range(len(data.station)):
-            curr_sta = data.station[i]
-            if curr_sta != prev_sta:
-                if prev_sta not in station_count_dic.keys():
-                    station_count_dic[prev_sta] = 1
-                else:
-                    station_count_dic[prev_sta] += 1
-                ind_end = i - 1
-                # create a new ChannelList object:
-                temp_sta = data.extract_subset_idx(ind_start, ind_end)
-                obstreestation = ObsTreeStation(
-                    temp_sta, prev_sta, str(station_count_dic[prev_sta])
-                )
-                obstreestation.meter_type = data.meter_type
-                logging.info(
-                    'Station added: ' + prev_sta + str(station_count_dic[prev_sta])
-                )
-                ind_start = i
-                self.appendRow(
-                    [obstreestation, QtGui.QStandardItem('0'), QtGui.QStandardItem('0')]
-                )
-            if i == len(data.station) - 1:
-                # enter last data line
-                ind_end = i
-                if curr_sta not in station_count_dic.keys():
-                    station_count_dic[curr_sta] = 1
-                else:
-                    station_count_dic[curr_sta] += 1
-
-                # create a new Loop-type object:
-                temp_sta = data.extract_subset_idx(ind_start, ind_end)
-                obstreestation = ObsTreeStation(
-                    temp_sta, curr_sta, str(station_count_dic[curr_sta])
-                )
-                obstreestation.meter_type = data.meter_type
-                logging.info(
-                    'Station added: '
-                    + curr_sta
-                    + '('
-                    + str(station_count_dic[curr_sta])
-                    + ')'
-                )
-                self.appendRow(
-                    [obstreestation, QtGui.QStandardItem('0'), QtGui.QStandardItem('0')]
-                )
-            prev_sta = curr_sta
-
-    def get_data_for_plot(self):
-        """
-        Retrieves data from loop that is used for plotting: occupations must be
-        checked, and there must be more than one occupation at a station.
-        :return: list of plot_data lists
-        """
-        unique_stations = set()
-        plot_data = []
-        for i in range(self.rowCount()):
-            station = self.child(i)
-            if self.child(i).data(role=Qt.CheckStateRole) == 2:
-                if (station.station_name, station.meter[0]) not in unique_stations:
-                    unique_stations.add(station.station_name)
-
-        for unique_station in unique_stations:
-            x, y, y_sd, t_sd = [], [], [], []
-            for i in range(self.rowCount()):
-                station = self.child(i)
-                if self.child(i).data(role=Qt.CheckStateRole) == 2:
-                    if station.station_name == unique_station:
-                        x.append(station.tmean())
-                        y.append(station.gmean())
-                        y_sd.append(station.original_sd)
-                        t_sd.append(station.t_stdev)
-            new_x, new_y, new_y_sd, new_t_sd = [], [], [], []
-            # -999's can occur when all samples at a station are unchecked
-            for idx, i in enumerate(x):
-                if i != -999:
-                    new_x.append(i)
-                    new_y.append(y[idx])
-                    new_y_sd.append(y_sd[idx])
-                    new_t_sd.append(t_sd[idx])
-            plot_data.append([new_x, new_y, unique_station, new_y_sd, new_t_sd])
-
-        # sort plot_data by initial x in each line
-        plot_data_sorted = []
-        exes = [x[0][0] for x in plot_data]
-
-        # get index of sorted list
-        idx = sorted(range(len(exes)), key=lambda m: exes[m])
-        for i in idx:
-            plot_data_sorted.append(plot_data[i])
-        return plot_data_sorted
-
-    def checked_stations(self):
-        """
-        Retrieves checked stations from loop; used for calculating delta-gs in
-        the "None" and "Continuous" drift options.
-        :return: list of checked stations
-        """
-        stations = []
-        for i in range(self.rowCount()):
-            station = self.child(i)
-            if self.child(i).data(role=Qt.CheckStateRole) == 2:
-                stations.append(station)
-        return stations
-
-    def stations(self):
-        """
-        """
-        stations = []
-        for i in range(self.rowCount()):
-            stations.append(self.child(i))
-        return stations
-
-    def n_unique_stations(self):
-        sn = [s.station_name for s in self.stations()]
-        return len(set(sn))
-
-    def deltas(self):
-        """
-        Called when 'Populate delta table... menu item is selected. Applies SD
-        settings in options.
-        """
-        deltas = []
-        table = self.delta_model
-        for i in range(table.rowCount()):
-            delta = table.data(table.index(i, 0), role=Qt.UserRole)
-            deltas.append(delta)
-            if self.parent().adjustment.adjustmentoptions.use_sigma_min:
-                delta.adj_sd = max(
-                    self.parent().adjustment.adjustmentoptions.sigma_min, delta.sd
-                )
-        return deltas
-
-    def to_json(self):
-        # Copy tares and stations from PyQt models to lists
-        tares = []
-        if self.tare_model is not None:
-            for i in range(self.tare_model.rowCount()):
-                ind = self.tare_model.createIndex(i, 0)
-                tares.append(self.tare_model.data(ind, Qt.UserRole))
-        stations_json = [s.to_json() for s in self.stations()]
-
-        self.delta_model = None
-        self.tare_model = None
-
-        return {
-            'checked': self.checkState(),
-            'delta_model': None,
-            'tare_model': None,
-            'stations': stations_json,
-            'tares': tares,
-            'name': self.name,
-            'drift_method': self.drift_method,
-            'drift_cont_method': self.drift_cont_method,
-            # If continuous model, also need to keep track of which type of model
-            'drift_cont_startend': self.drift_cont_startend,  # behavior at start/end. 0: extrapolate, 1: constant
-            'drift_netadj_method': self.drift_netadj_method,  # If netadj method, keep track of polynomial degree
-            'meter': self.meter,  # Meter S/N, for the case where multiple meters are calibrated
-            'comment': self.comment,
-            'oper': self.oper,
-        }
-
-
-class ObsTreeSurvey(ObsTreeItem):
+class ObsTreeSurvey(ObsTreeItemBase):
     """
     A survey object is a little different than Loop or Station objects in that
     it doesn't have a corresponding data object. All relevant info is in
@@ -711,7 +105,7 @@ class ObsTreeSurvey(ObsTreeItem):
             sd.type = delta['type']
             try:
                 sd.assigned_dg = delta['assigned_dg']
-            except:
+            except Exception:
                 pass
             try:
                 sd.sta1 = delta['sta1']
@@ -1276,7 +670,7 @@ class ObsTreeSurvey(ObsTreeItem):
                 write_out_files='n',
             )
         except InversionError:
-            show_message(InversionError.message, "Inversion Error")
+            show_message(InversionError, "Inversion Error")
         self.match_inversion_results('numpy', cal_dic)
 
     def match_inversion_results(self, inversion_type, cal_dic=None):
@@ -1446,316 +840,3 @@ class ObsTreeSurvey(ObsTreeItem):
                             "GSadjust error",
                         )
         return True
-
-
-class ObsTreeModel(QtGui.QStandardItemModel):
-    """
-    Tree model that shows station name, date, and average g value.
-
-    The model is populated by appending stations to loops, and loops to surveys.
-    The parent-child relationship is not explicitly stored.
-    """
-
-    # TODO: Implement drag and drop.
-    def __init__(self):
-        super(ObsTreeModel, self).__init__()
-        self.setColumnCount(3)
-        self.setHorizontalHeaderLabels(['Name', 'Date', 'g (\u00b5Gal)'])
-        self.station_coords = None
-
-    #
-    # signal_refresh_view = QtCore.pyqtSignal()
-    signal_name_changed = QtCore.pyqtSignal()
-    signal_delta_update_required = QtCore.pyqtSignal()
-
-    def columnCount(self, QModelIndex_parent=None, *args, **kwargs):
-        return 3
-
-    def flags(self, QModelIndex):
-        if not QModelIndex.isValid():
-            return Qt.NoItemFlags
-        return (
-            Qt.ItemIsEnabled
-            | Qt.ItemIsSelectable
-            | Qt.ItemIsEditable
-            | Qt.ItemIsUserCheckable
-        )
-
-    def data(self, index, role=Qt.DisplayRole):
-        if index.model() is not None:
-            column = index.column()
-            if role == Qt.DisplayRole:
-
-                if column > 0:
-                    m = index.model().itemFromIndex(index.sibling(index.row(), 0))
-                else:
-                    m = index.model().itemFromIndex(index)
-                try:  # Was getting "AttributeError: 'QStandardItem' object has no attribute 'display_column_map'"
-                    # after deleting a survey
-                    fn, *args = m.display_column_map.get(
-                        column, (format_numeric_column, column)
-                    )
-                    return fn(*args)
-                except AttributeError:
-                    return ''
-
-            elif role == Qt.CheckStateRole:
-                if column == 0:
-                    m = index.model().itemFromIndex(index)
-                    return m.checkState()
-            elif role == Qt.ToolTipRole:
-                m = index.model().itemFromIndex(index)
-                try:
-                    return m.tooltip
-                except AttributeError:
-                    return ''
-
-    def setData(self, index, value, role):
-        if role == Qt.CheckStateRole and index.column() == 0:
-            m = index.model().itemFromIndex(index)
-            m.setCheckState(value)
-            if type(m) is ObsTreeLoop:
-                self.signal_delta_update_required.emit()
-            self.dataChanged.emit(index, index)
-            # # self.layoutChanged.emit()
-            # self.signal_refresh_view.emit()
-            return True
-
-        if role == Qt.EditRole:
-            if index.isValid() and index.column() == 0:
-                item = index.model().itemFromIndex(index)
-                return {
-                    ObsTreeStation: self._handler_edit_ObsTreeStation,
-                    ObsTreeLoop: self._handler_edit_ObsTreeLoop,
-                    ObsTreeSurvey: self._handler_edit_ObsTreeSurvey,
-                }[type(item)](item, value)
-
-    def _handler_edit_ObsTreeStation(self, item, value):
-        """
-        Handle editing of ObsTreeStation objects (renaming).
-        """
-        from gui_objects import rename_dialog
-
-        old_name = item.station_name
-        new_name = str(value)
-        if new_name is not item.station_name:
-            rename_type = rename_dialog(old_name, new_name)
-            if rename_type == 'Loop':
-                loop = item.parent()
-                loop.rename(old_name, new_name)
-
-            if rename_type == 'Survey':
-                loop = item.parent()
-                survey = loop.parent()
-                survey.rename(old_name, new_name)
-
-            if rename_type == 'Campaign':
-                campaign = item.model().invisibleRootItem()
-                for i in range(campaign.rowCount()):
-                    survey = campaign.child(i, 0)
-                    survey.rename(old_name, new_name)
-
-            if rename_type == 'Station':
-                item.station_name = new_name
-                for i in range(len(item.station)):
-                    item.station[i] = new_name
-
-            logging.info(
-                'Stations renamed from {} to {} in {}'.format(
-                    old_name, new_name, rename_type
-                )
-            )
-            self.signal_name_changed.emit()
-        return True
-
-    def _handler_edit_ObsTreeLoop(self, item, value):
-        """
-        Handle editing of ObsTreeLoop objects (renaming).
-        """
-        new_name = str(value)
-        old_name = item.name
-        logging.info('Loop renamed from {} to {}'.format(old_name, new_name))
-        item.name = new_name
-        return True
-
-    def _handler_edit_ObsTreeSurvey(self, item, value):
-        new_name = str(value)
-        try:
-            name_as_date = dt.datetime.strptime(new_name, '%Y-%m-%d')
-            old_name = name_as_date
-            logging.info('Loop renamed from {} to {}'.format(old_name, new_name))
-            item.name = new_name
-            return True
-        except Exception as e:
-            return False
-
-    def checked_surveys(self):
-        """
-        Retrieves checked data from loop; used for calculating delta-gs in the
-        "None" and "Continuous" drift options.
-        :return: list of checked stations
-        """
-        data = []
-        for i in range(self.invisibleRootItem().rowCount()):
-            if self.invisibleRootItem().child(i).checkState() == 2:
-                data.append(self.invisibleRootItem().child(i))
-        return data
-
-    def unique_meters(self):
-        meters = []
-        for survey in self.checked_surveys():
-            meters += survey.unique_meters
-        return list(set(meters))
-
-    def get_cal_coeffs(self):
-        cal_dict = {}
-        for m in self.unique_meters():
-            cal_dict[m] = []
-        for survey in self.checked_surveys():
-            if survey.adjustment.adjustmentoptions.cal_coeff:
-                for k, v in survey.adjustment.adjustmentresults.cal_dic.items():
-                    cal_dict[k].append((survey.name, v[0], v[1]))
-            elif survey.adjustment.adjustmentoptions.specify_cal_coeff:
-                for k, v in survey.adjustment.adjustmentoptions.meter_cal_dict.items():
-                    cal_dict[k].append((survey.name, v, 0))
-        return cal_dict
-
-    def checkState(self, index):
-        if index.column > 0:
-            m = index.model().itemFromIndex(index.sibling(index.row(), 0))
-        else:
-            m = index.model().itemFromIndex(index)
-        return m.checkState()
-
-    def insertRows(self, station, position, rows=1, index=QtCore.QModelIndex()):
-        self.beginInsertRows(index, position, position + rows - 1)
-        self.appendRow(station)
-        self.endInsertRows()
-
-    def load_workspace(self, fname):
-        """
-        Load previously-save workspace. Need to recreate PyQt models.
-
-        Importantly, there are two types of deltas: those on the Drift tab and
-        those on the network adjustment tab. deltas are COPIED from one to
-        another by discrete menu commands, it's not automatic.
-
-        The workflow:
-        1) Create obstreesurvey object so we have somewhere to store loop and
-            station objects.
-        2) deltas on the drift tab are created strictly from the station objects
-            and specified options (drift correction method, etc.). We don't store
-            a corresponding delta object in the saved workspace/json.
-        3) deltas on the network adjustment tab have additional information that
-            must be stored in the saved workspace/json (checked state, std. dev.
-            for adj., etc.). When loading a workspace, this delta is INDEPENDENT of
-            the delta stored in the delta table on the Drift tab. The Net Adj. tab
-            delta table is re-created later, not here, because Roman-method deltas
-            depend on the drift-tab deltas.
-        :param fname:
-        :return: (ObsTreeSurvey, delta_table, coords)
-        """
-        logging.info("Workspace loaded: " + fname)
-        delta_tables, obstreesurveys = [], []
-        coords, surveys = None, None
-        with open(fname, "r") as f:
-            data = jsons.load(json.load(f))
-            if all(isinstance(x, ObsTreeSurvey) for x in data):
-                surveys = data
-            elif len(data) > 1:
-                coords = data[1]
-                surveys = data[0]
-        # Populate PyQt objects. First, create obstreesurvey
-        for survey in surveys:
-            obstreesurvey = ObsTreeSurvey.from_json(survey)
-            for loop in survey['loops']:
-                obstreeloop = ObsTreeLoop.from_json(loop)
-                for station in loop['stations']:
-                    if (
-                        'station_name' in station
-                    ):  # Sometimes blank stations are generated, not sure why?
-                        temp_station = tempStation(station)
-                        obstreestation = ObsTreeStation(
-                            temp_station,
-                            temp_station.station_name,
-                            temp_station.station_count,
-                        )
-                        if type(obstreestation.t[0]) == dt.datetime:
-                            obstreestation.t = [date2num(i) for i in obstreestation.t]
-                        obstreeloop.appendRow(
-                            [
-                                obstreestation,
-                                QtGui.QStandardItem('0'),
-                                QtGui.QStandardItem('0'),
-                            ]
-                        )
-                # stations is provided by loop.stations(), it doesn't need to be stored separately in the ObsTreeLoop
-                # object
-                if hasattr(obstreeloop, 'stations'):
-                    del obstreeloop.stations
-                obstreesurvey.appendRow(
-                    [obstreeloop, QtGui.QStandardItem('0'), QtGui.QStandardItem('0')]
-                )
-            obstreesurveys.append(obstreesurvey)
-            delta_tables.append(obstreesurvey.deltas)
-
-        return (obstreesurveys, delta_tables, coords)
-
-    def datums(self):
-        datum_list = []
-        for i in range(self.rowCount()):
-            obstreesurvey = self.itemFromIndex(self.index(i, 0))
-            for ii in range(obstreesurvey.datum_model.rowCount()):
-                idx = obstreesurvey.datum_model.index(ii, 0)
-                datum = obstreesurvey.datum_model.data(idx, role=Qt.UserRole)
-                datum_list.append(datum.station)
-        return list(set(datum_list))
-
-    def surveys(self):
-        survey_list = []
-        for i in range(self.rowCount()):
-            obstreesurvey = self.itemFromIndex(self.index(i, 0))
-            survey_list.append(obstreesurvey)
-        return survey_list
-
-    def deltas(self):
-        delta_list = []
-        for obstreesurvey in self.surveys():
-            delta_list += obstreesurvey.delta_list()
-        return delta_list
-
-    def stations(self):
-        station_list = []
-        for obstreesurvey in self.surveys():
-            for obstreeloop in obstreesurvey.loops():
-                station_list += obstreeloop.stations()
-        return list(set(station_list))
-
-    def resetStationAsd(self):
-        for survey in self.surveys():
-            for loop in survey.loops():
-                for station in loop.stations():
-                    station.assigned_sd = None
-
-    def save_workspace(self, fname):
-        surveys = []
-        for i in range(self.rowCount()):
-            obstreesurvey = self.itemFromIndex(self.index(i, 0))
-            surveys.append(obstreesurvey)
-        workspace_data = [surveys, self.station_coords]
-        if fname[-4:] != '.gsa':
-            fname += '.gsa'
-        with open(fname, "w") as f:
-            json.dump(jsons.dump(workspace_data), f)
-        logging.info('Saving JSON workspace to {}'.format(fname))
-        return fname
-
-
-def survey_serializer(obj, cls, **kwargs):
-    """
-    Handle serialization of ObsTreeSurvey via .to_json() method.
-    """
-    return obj.to_json()
-
-
-jsons.set_serializer(survey_serializer, ObsTreeSurvey)
